Avoid unnecessary error messages from hwcomposer under crosvm
When the vnc server isn't started, there is no vsock server on the
host for the hwcomposer to connect to, so it fails its connection
attempt, logs it and continues working without sending the composed
frames to the host. Instead, it shouldn't even try to connect when the
server is known to not be available.

Test: run locally
Bug: 132114593
Change-Id: I3ec2bc0b2e97ba73c87ccbced710480d42846872
(cherry picked from commit 1859cccc9d2c56fe5dce7213f680621a044b97d5)