Mild hack: copy bpttool into cuttlefish_common
bpttool comes from here:
https://android.googlesource.com/platform/system/tools/bpt/+/refs/heads/master/bpttool

Copying the one file is easier than fixing the manifest (see b/78137118)

File renamed from bpttool to cf_bpttool to avoid build conflicts when
the real project is added to the tree.

Alternatively, the code could be reimplemented later in C++ to avoid a
dependency on python in the host image.

Bug: 133432409
Change-Id: Ia1036ff79e7cd5ec9be2f5ddce7bc67c6dcb1c04