Crosvm HWComposer HAL gets configuration from configuration server
Instead of from screen server.

As a side effect, this eliminates a race in which the constructor of
the FrameBuffer takes too long to connect to the screen server (to
obtain the screen config) and then broadcaster thread exits because
there is (still) no connection to the screen server. With this change
the constructor connects to the configuration server while the
broadcaster thread connects to the screen server.

Bug: 134972437
Test: run locally
Change-Id: I06dc3f51b21ef4e5192b0bb9cec7405bcf83dcfe