Eliminate code duplication from hwcomposer
Move all common code to guest/hals/hwcomposer/common, which includes
the implementation of cpu-based compositions (vsoc_composer.* files
are renamed to cpu_composer.*).
Encapsulate the differences in classes implementing the contract
defined in the ScreenView abstract class (VsocketScreenView for crosvm
and VSoCScreenView for qemu).
Fix formatting of some files.

Bug: 135628605
Test: run locally with qemu & crosvm
Change-Id: Iae3c8a31155f935a4ee60ef1e16c934ad4878587