Support downloading the target files zip.
This has all the component parts for building super images. Enabled
only with a flag since it's a significantly larger download than the
img-*.zip file, though the GCE to Build API throughput is pretty good.

At the moment it is not possible to distinguish between which target
files are for the default image and which are for the system image. That
will be addressed in the next CL.

Bug: 135293952
Test: Run fetch_cvd -download_target_files_zip
Change-Id: I037cd1f7c93f1f3ff44d827574dd35935f828288