Make stop_cvd (more) synchronous
Instead of claiming success when the SIGKILL signal is sent, the
stopping code now waits for each subprocess to actually die before
claiming success.
Because the only process that can wait for host processes is the
launcher (their parent) and because the launcher won't be able to the
waiting if it sends a SIGKILL to its own group, every subprocess
started by the launcher through the process monitor is the head of its
own process group. On stop, the launcher sends SIGKILL to each
subprocess' process group and waits on each of them before reporting
success to stop_cvd and exiting.
It's still technically possible for the launcher or some of its
grandchildren to still be running after stop_cvd exits, but it's less
likely now:
- The launcher calls write on a socket (shouldn't block) and
exits immediately, so it's likely gone by the time stop_cvd reads from
the socket.
- The only grandchildren of the launcher seem to be processes started
by crosvm, which are stopped first by calling vm_manager->Stop().

Bug: 140224512
Test: launch_cvd -daemon; stop_cvd; launch_cvd
Change-Id: I9a912c0bae4f0873fec9bc868f0a7f71d1b38e2a