Set the generated ramdisk's location in the config
The CuttlefishConfig object was being saved to disk and then modified
in-memory afterwards. This worked when launch_cvd carried a lot more
logic, but with the assemble_cvd and run_cvd split exposed this issue as
assemble_cvd has to write the config to disk and run_cvd picks it up.

Test: launch_cvd, touch empty && launch_cvd --initramfs_path=empty
Test: hridya@ validated aosp/1132046 does work with this change
Bug: 142142809
Change-Id: I16b571d8dbe3b24c45f8c43047d4a6424de3f725