Support img.zip files without system images.
Mixed system image builds will now always download the target files zip
in addition to the image zip, which has more than enough files to
support remixing the super image.

This is a simple solution to the problem but adds a nontrivial download
cost, so there is a lot of room for optimization here.

Bug: 141721921
Test: fetch_cvd -default_build=5925895 -system_build=5926038
Change-Id: I573b40483ba7730e89c4062366e761399666728d